chess_candidates_2026/requirements_chess.txt
python-chess>=1.10.0
requests>=2.32.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
lightgbm>=4.0.0
scikit-learn>=1.4.0
pyyaml>=6.0.0
plotly>=5.24.0
streamlit>=1.40.0
